+++ generators/index.ts
@@ -32,6 +32,7 @@ export { clientProviderGenerator } from './client/client-provider.generator.js';
 export { clientInterceptorGenerator } from './client/client-interceptor.generator.js';
 export { clientGuardGenerator } from './client/client-guard.generator.js';
 export { clientInterfaceDefinitionsGenerator } from './client/client-interface-definitions.generator.js';
+export { packageServiceGenerator } from './client/package-service.generator.js';
 // Server generators
 export { serverGenerator } from './server/server.generator.js';
 export { serverPackageGenerator } from './server/server-package.generator.js';